requirements.txt
python-vlc
pafy
youtube-dl